Bugfix 474: Define socklen_t when undefined. Thanks Frank Denis

+++ drill/drill_util.c
@@ -14,7 +14,7 @@
 #include <errno.h>
 
 static int
-read_line(FILE *input, char *line, size_t len)
+read_line(FILE *input, char *line, int len)
 {
 	int i;
